recovery: add a "rescue mode" (SD-boot recovery managing an eMMC install)
Recoveries which boot off SD card, but manage an eMMC install are
indispensable for recovering from certain catastrophes.  It's useful to
be able to create such a "rescue" recovery image without having to
modify the recovery initramfs; to facilitate this, check for a flag
under /boot on startup and operate on eMMC regardless of boot location
if it's present.

Change-Id: I95b5523a1371833b405c17dfbdc7b6475c82fad0