recovery: add support for command files in /boot
Android recoveries accept a list of commands to perform on startup in
/cache/recovery/command; while this is a good location for an installed
system to communicate with a recovery, it doesn't work for creating
special-purpose script-driven recovery images (for initial install or
repartitioning, for example).

As a solution, introduce a command file in /boot (we have to be able to
ship files in /boot to have a recovery image anyway, so this is a
natural place to put it), and copy the arguments from that file to the
bootloader control block on recovery startup so that they'll be picked
up.

Change-Id: Ic06850d02ef7d6f59f6ae99f3d0eef79b82e1857