configure_vold.sh: use symlink instead of property to detect SD installs
In recovery rescue mode, we want everything set up as if it were an eMMC
install, even though we booted off SD card.  Since the
ro.boot.bootdevice property can't be changed for rescue mode, we need to
use the /dev/block/by-name symlinks to determine whether we want to
configure vold for an SD card install instead.

Change-Id: I2a054f81521f8ba09335adfe5084bb6b1dc6454d