Update bootloader to fix arguments when user selects recovery boot
The bootloader was passing in --wipe_data_ui to the recovery via the BCB
when a boot-recovery menu entry was selected; with the recent change to
the recovery to honor arguments in the BCB, this argument was getting
picked up (we're very fortunate this appears to be specific to the BN
recovery and isn't honored by other recoveries).  Update the bootloader
to avoid this.

Built from commit 502626e ("menu: don't command a data wipe when
selecting boot-to-recovery options") in the encore_cyanoboot tree.

Change-Id: Iad99ccbaceb60225990b57b08b929215facd7e8d