Detect and support a combined data/media partition layout
We'd like to move towards a combined data/media partition layout for
future releases of Android -- devices with 1 GB /data will struggle with
ART enabled, while a 1 GB /emmc isn't terribly useful in this day and
age.  To this end, we'd like to introduce the option for users to
repartition their devices during the 4.4/CM11 release cycle.

For this to work, we need to support both traditional (separate /data
and /emmc) and unified layouts.  This requires changes to MountService
to allow configuring storage properties at runtime; with that in place,
we need only detect unified layouts, set the appropriate properties, and
start the FUSE sdcard emulation service when necessary.  (There's the
minor complication that datamedia isn't actually on /data/media for SD
card installs, so we have to detect that case and handle it separately.)

Change-Id: I80d89aac59bcb2018ef7ad4920d87941c3cab63c