Fix restoring gapps on flashing.
It almost works. Because of our severe memory limitations
and an inability to automatically mount /sdcard in recovery.
After you have booted into recovery you need to run:
$ adb shell
From within the shell run the following commands (but not the #)
This gives the gapps backup script a place to put the installed gapps.
Then you can go ahead and flash the rom. You won't lose your gapps.

Change-Id: I42f547361f9919d34549617957cc4b1dda2ab663