Fix backup of gapps on flash.
We no longer need to do the adb mounting. It is done automatically.
It requires an external sdcard. There may be problems if it is
set as adopted storage since marshmallow encrypts that. It does
work if the sdcard is simple removable storage.
There are a couple of error messages in the recovery.log relating
to mounting sdcard but it mounts and restores gapps just fine.

Change-Id: I4b2438b3e74435f9dcbe1db091cf744fcf9b5dae